@@ -1,10 +1,10 @@
 name: peercoin
 description: A new Peercoin wallet.
 
-version: 1.2.2+134
+version: 1.2.3+135
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   
 publish_to: none
 
@@ -55,7 +55,7 @@ dependencies:
   app_bar_with_search_switch: ^1.5.4
   decimal: ^2.3.3
   characters: ^1.2.0
-  coinlib_flutter: ^1.0.0
+  coinlib_flutter: ^2.0.0-rc.3
 
 dev_dependencies:
   integration_test:
@@ -68,7 +68,7 @@ dev_dependencies:
   build_runner: ^2.0.4
   flutter_launcher_icons: ^0.11.0
   flutter_native_splash: ^2.2.7
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.0
 
 flutter_icons:
   android: "launcher_icon"
