@@ -38,10 +38,10 @@ class WalletListScreen extends StatefulWidget {
   State<WalletListScreen> createState() => _WalletListScreenState();
 
   const WalletListScreen({
-    Key? key,
+    super.key,
     this.fromColdStart = false,
     this.walletToOpenDirectly = '',
-  }) : super(key: key);
+  });
 }
 
 class _WalletListScreenState extends State<WalletListScreen>
@@ -452,14 +452,29 @@ class _WalletListScreenState extends State<WalletListScreen>
                                           );
                                         },
                                         child: ListTile(
-                                          leading: CircleAvatar(
-                                            backgroundColor: Colors.white,
-                                            child: Image.asset(
-                                              AvailableCoins.getSpecificCoin(
-                                                wallet.name,
-                                              ).iconPath,
-                                              width: 20,
-                                            ),
+                                          leading: Stack(
+                                            children: [
+                                              CircleAvatar(
+                                                backgroundColor: Colors.white,
+                                                child: Image.asset(
+                                                  AvailableCoins
+                                                      .getSpecificCoin(
+                                                    wallet.name,
+                                                  ).iconPath,
+                                                  width: 20,
+                                                ),
+                                              ),
+                                              if (wallet.watchOnly)
+                                                const Positioned(
+                                                  right: 0,
+                                                  bottom: 0,
+                                                  child: Icon(
+                                                    Icons.visibility,
+                                                    size: 16,
+                                                    color: Colors.black,
+                                                  ),
+                                                ),
+                                            ],
                                           ),
                                           title: Text(
                                             wallet.title,
